# WIP: Reduce amount of std::map lookups performed during a sim