--- a/OpenSim/Actuators/Millard2012EquilibriumMuscle.cpp
+++ b/OpenSim/Actuators/Millard2012EquilibriumMuscle.cpp
@@ -348,8 +348,8 @@ setActivation(SimTK::State& s, double activation) const
         setStateVariableValue(s, STATE_ACTIVATION_NAME,
                               getActivationModel().clampActivation(activation));
     }
-    this->markCacheVariableInvalid(s, this->velInfoCV);
-    this->markCacheVariableInvalid(s, this->dynamicsInfoCV);
+    this->markCacheVariableInvalid(s, _velInfoCV);
+    this->markCacheVariableInvalid(s, _dynamicsInfoCV);
 }
 
 void Millard2012EquilibriumMuscle::setDefaultFiberLength(double fiberLength)
@@ -389,9 +389,9 @@ setFiberLength(SimTK::State& s, double fiberLength) const
     if (!get_ignore_tendon_compliance()) {
         setStateVariableValue(s, STATE_FIBER_LENGTH_NAME,
                               clampFiberLength(fiberLength));
-        this->markCacheVariableInvalid(s, this->lengthInfoCV);
-        this->markCacheVariableInvalid(s, this->velInfoCV);
-        this->markCacheVariableInvalid(s, this->dynamicsInfoCV);
+        this->markCacheVariableInvalid(s, _lengthInfoCV);
+        this->markCacheVariableInvalid(s, _velInfoCV);
+        this->markCacheVariableInvalid(s, _dynamicsInfoCV);
     }
 }