--- a/OpenSim/Common/CommonUtilities.cpp
+++ b/OpenSim/Common/CommonUtilities.cpp
@@ -23,11 +23,16 @@
 
 #include "CommonUtilities.h"
 
+#include "PiecewiseLinearFunction.h"
+#include "STOFileAdapter.h"
+#include "TimeSeriesTable.h"
 #include <chrono>
 #include <ctime>
 #include <iomanip>
 #include <memory>
 #include <sstream>
+
+#include <SimTKcommon/internal/Pathname.h>
 
 std::string OpenSim::getFormattedDateTime(
         bool appendMicroseconds, std::string format) {
@@ -62,4 +67,118 @@ std::string OpenSim::getFormattedDateTime(
         ss << '.' << std::setfill('0') << std::setw(6) << microsec.count();
     }
     return ss.str();
+}
+
+SimTK::Vector OpenSim::createVectorLinspace(
+        int length, double start, double end) {
+    SimTK::Vector v(length);
+    for (int i = 0; i < length; ++i) {
+        v[i] = start + i * (end - start) / (length - 1);
+    }
+    return v;
+}
+
+SimTK::Vector OpenSim::createVector(
+        std::initializer_list<SimTK::Real> elements) {
+    return SimTK::Vector((int)elements.size(), elements.begin());
+}
+
+SimTK::Vector OpenSim::interpolate(const SimTK::Vector& x,
+        const SimTK::Vector& y, const SimTK::Vector& newX,
+        const bool ignoreNaNs) {
+
+    OPENSIM_THROW_IF(x.size() != y.size(), Exception,
+            "Expected size of x to equal size of y, but size of x "
+            "is {} and size of y is {}.",
+            x.size(), y.size());
+
+    // Create vectors of non-NaN values if user set 'ignoreNaNs' argument to
+    // 'true', otherwise throw an exception. If no NaN's are present in the
+    // provided data vectors, the '*_no_nans' variables below will contain
+    // the original data vector values.
+    std::vector<double> x_no_nans;
+    std::vector<double> y_no_nans;
+    for (int i = 0; i < x.size(); ++i) {
+
+        bool shouldNotPushBack =
+                (SimTK::isNaN(x[i]) || SimTK::isNaN(y[i])) && ignoreNaNs;
+        if (!shouldNotPushBack) {
+            x_no_nans.push_back(x[i]);
+            y_no_nans.push_back(y[i]);
+        }
+    }
+
+    OPENSIM_THROW_IF(x_no_nans.empty(), Exception,
+            "Input vectors are empty (perhaps after removing NaNs).");
+
+    PiecewiseLinearFunction function(
+            (int)x_no_nans.size(), &x_no_nans[0], &y_no_nans[0]);
+    SimTK::Vector newY(newX.size(), SimTK::NaN);
+    for (int i = 0; i < newX.size(); ++i) {
+        const auto& newXi = newX[i];
+        if (x_no_nans[0] <= newXi && newXi <= x_no_nans[x_no_nans.size() - 1])
+            newY[i] = function.calcValue(SimTK::Vector(1, newXi));
+    }
+    return newY;
+}
+
+std::string OpenSim::getAbsolutePathnameFromXMLDocument(
+        const std::string& documentFileName,
+        const std::string& pathnameRelativeToDocument) {
+    // Get the directory containing the XML file.
+    std::string directory;
+    bool dontApplySearchPath;
+    std::string fileName, extension;
+    SimTK::Pathname::deconstructPathname(documentFileName, dontApplySearchPath,
+                                         directory, fileName, extension);
+    return SimTK::Pathname::getAbsolutePathnameUsingSpecifiedWorkingDirectory(
+            directory, pathnameRelativeToDocument);
+}
+
+SimTK::Real OpenSim::solveBisection(
+        std::function<SimTK::Real(const SimTK::Real&)> calcResidual,
+        SimTK::Real left, SimTK::Real right, const SimTK::Real& tolerance,
+        int maxIterations) {
+    SimTK::Real midpoint = left;
+
+    OPENSIM_THROW_IF(maxIterations < 0, Exception,
+            "Expected maxIterations to be positive, but got {}.",
+            maxIterations);
+
+    const bool sameSign = calcResidual(left) * calcResidual(right) >= 0;
+    if (sameSign && Logger::shouldLog(Logger::Level::Debug)) {
+        const auto x = createVectorLinspace(1000, left, right);
+        TimeSeriesTable table;
+        table.setColumnLabels({"residual"});
+        SimTK::RowVector row(1);
+        for (int i = 0; i < x.nrow(); ++i) {
+            row[0] = calcResidual(x[i]);
+            table.appendRow(x[i], row);
+        }
+        STOFileAdapter::write(
+                table, fmt::format("solveBisection_residual_{}.sto",
+                               getFormattedDateTime()));
+    }
+    OPENSIM_THROW_IF(sameSign, Exception,
+            "Function has same sign at bounds of {} and {}.", left, right);
+
+    SimTK::Real residualMidpoint;
+    SimTK::Real residualLeft = calcResidual(left);
+    int iterCount = 0;
+    while (iterCount < maxIterations && (right - left) > tolerance) {
+        midpoint = 0.5 * (left + right);
+        residualMidpoint = calcResidual(midpoint);
+        if (residualMidpoint * residualLeft < 0) {
+            // The solution is to the left of the current midpoint.
+            right = midpoint;
+        } else {
+            left = midpoint;
+            residualLeft = calcResidual(left);
+        }
+        ++iterCount;
+    }
+    if (iterCount == maxIterations) {
+        log_warn("Bisection reached max iterations at x = {}.", midpoint);
+    }
+    return midpoint;
 }