--- a/OpenSim/Common/CommonUtilities.h
+++ b/OpenSim/Common/CommonUtilities.h
@@ -24,7 +24,9 @@
  * -------------------------------------------------------------------------- */
 
 #include "osimCommonDLL.h"
+#include <functional>
 #include <iostream>
+#include <memory>
 
 #include <SimTKcommon/internal/BigMatrix.h>