This PR covers reducing the amount of runtime lookups done by the various models, muscles, etc. in OpenSim at runtime.

# Overview

Runtime `std::map` lookups can happen *a lot* in a simulation. The reason is probably because many concrete simulation/model implementations derive from `Component`, which has its own storage model for storing + caching things in the model graph. Derived classes are incentivized to perform lookups using the base `Component` class, rather than handling lookups themselves, because the code + datastructures are already there, so it's more convenient than caching the lookup indexes locally etc.

An example of where this happens: a class may want runtime access to some cached variable of type `T` in the current model being simulated. To do that, a common pattern is to call something like `this->getModel().getCacheVariableValue<T>(st, "foo");`. This ultimately breaks down into:

- Get the current model (a pointer lookup + dereference) 
- Lookup the cache variable in a `Component::_namedCacheVariableInfo`, which is a `std::map`
- Get the `CacheEntryIndex` from the `CacheInfo` held in the `std::map`
- Get the `SimTK::System` the `Component` is attached to
- Use the `CacheEntryIndex` to get the cache entry from the `SimTK::System`
- Dereference the resulting entry
- Downcast the entry to `T` with a `dynamic_cast`

Ignoring the details around pointer types, memory layout, etc. cycle estimation (valgrind) measurements on (e.g.) `ToyDropLanding` show results like:

| Function Name | Called # | Incl. % | Self % | 
| - | - | - | - |
| isCacheVariableValid | 12 686 729 | 5.24 | 1.26 |
| setCacheVariableValue<double> | 5 589 111 | 2.28 | 0.58 |
| getCacheVariableValue<double> | 6 457 708 | 2.13 | 0.58 |
| markCacheVariableValid | 3 964 799 | 1.52 | 0.29 |

Not a massive deal--it's ultimately < 12 % of perf--but there might be some low-hanging fruit in terms of perf optimizations we can roll out to reduce the overhead of those functions. However, these higher-level lookup calls ultimately result in the profiler measuring a bunch of overhead in  related, lower-level, operations (string storage, map lookups, downcasting):

| Function Name | Called # | Incl. % | Self % | 
| - | - | - | - |
| std::_Rb_tree<...CacheInfo...> (lookups) | 39 008 221 | 10.42 | 5.23 |
| std::basic_string<char> (ctor, key copies?) | 58 939 682 | 3.75 | 2.36 |
| __dynamic_cast (downcasts) | 51 315 491 | 3.57 | 1.28 |

# Other Points

- @tgeijten has already done a bunch of legwork on this: https://github.com/tgeijten/opensim32/commit/56f825c4d014826ee073d4025a257ce6e58ebd44 . I separately measured similar perf problems before I knew about this.

- Additionally, I performed extra measurements (callgrind simulations) and looked into theories around cache misses etc. Looking into it, the biggest source of L1 misses:

  - ~3.1 %: `Rb_tree ` lookups
  - ~2.9 %: `Property<double>::getValue(int)`
  - ~2 %: `realizeMeasureTimeVirtual(SimTK::State&)`

- Which is to say that there's nothing *obviously* evicting and reading L1 in one localized place (e.g. a loop that hits a large array of data, so all L1 fetches happen in that function). The L1 profile indicates that L1 misses are happening in *many* places all over OpenSim, which is *probably* because there's a huge amount of method nesting, getter nesting, abstraction, etc. that happens all over the place, rather than there being one obvious loop somewhere. 

- L3 misses:

  - 30 % `_dl_relocate_object`
  - 24 % `_dl_lookup_symbol_x`
  - 7 % `do_lookup_x`

- Which indicates that the application has to fetch its functions from main memory, probably at loadup. Assuming that, once the application is loaded, it spends most of its time in L1-L3. Other functions in the L3 simulation all look like object destructors etc. (also unsurprising). 

- To confirm this, I ran `valgrind --tool=massif` on the application, which recorded that peak heap usage during the simulation is ~110 MB, with 80 MB being `mmaped` (probably libraries, which will page-miss, and L3 miss, during the instruction fetch), 15 MB being from `operator new` (probably OpenSim/Simbody) and 8 MB being from `malloc`. My CPU (Ryzen 1600) has 16 MB of L3, and all of the memory will not be used in the core loop, so it's possible that OpenSim can reside entirely in L3 for the `ToyDropLanding` sim.

# ToDo (from conculsions)

- Almost all perf hits are for cache variable lookups. There are other `std::map`s in `Component`, but the biggest perf gains for the smallest amount of work should focus on just optimizing any of the `(get|upd|set)*Cache*` methods