@adamkewley 2020-06-15
So, ignoring the merge bs, etc. `d4bb192` is essentially just a cleaned-up version of the original code that also optimizes out a few unnecessary copies during `CacheEntry` initialization. Passes tests and nothing particularly noteworthy in terms of perf:

repeats = 32
|                  Test Name | master [secs] | σ [secs] | branch [secs] | σ [secs] | Speedup |
| -------------------------- | ------------- | -------- | ------------- | -------- | ------- |
|                   Gait2354 |          0.26 |     0.00 |          0.26 |     0.00 |    1.00 |
|             ToyDropLanding |         15.18 |     0.04 |         15.12 |     0.04 |    1.00 |
|   ToyDropLanding_nomuscles |          0.44 |     0.00 |          0.44 |     0.00 |    1.01 |
|            passive_dynamic |          5.01 |     0.01 |          5.00 |     0.02 |    1.00 |
| passive_dynamic_noanalysis |          3.18 |     0.02 |          3.19 |     0.02 |    1.00 |
|                      Arm26 |          0.31 |     0.00 |          0.30 |     0.00 |    1.01 |
|             RajagopalModel |         16.59 |     0.14 |         16.02 |     0.05 |    1.04 |