@adamkewley 2020-06-17
The refactored version basically has no effect on perf overall. Any changes in speed are basically noise:

repeats = 64
|                  Test Name | master [secs] | σ [secs] | branch [secs] | σ [secs] | Speedup |
| -------------------------- | ------------- | -------- | ------------- | -------- | ------- |
|                   Gait2354 |          0.26 |     0.00 |          0.27 |     0.00 |    0.98 |
|             ToyDropLanding |         15.38 |     0.10 |         15.60 |     0.08 |    0.99 |
|   ToyDropLanding_nomuscles |          0.45 |     0.00 |          0.45 |     0.00 |    1.00 |
|            passive_dynamic |          5.06 |     0.02 |          5.07 |     0.01 |    1.00 |
| passive_dynamic_noanalysis |          3.19 |     0.01 |          3.19 |     0.02 |    1.00 |
|                      Arm26 |          0.31 |     0.00 |          0.31 |     0.00 |    1.03 |
|             RajagopalModel |         16.23 |     0.07 |         16.25 |     0.14 |    1.00 |

Which is to be expected, because the sequence of operations in this refactor *should* be exactly the same as the pre-refactor.