@adamkewley 2020-06-17
So the next thing to do is decide on how to "skip" the `CacheVariable` lookup (an expensive `std::map` lookup), by providing callers with a direct handle to (e.g.) the underlying index. 

This all sounds easy and sane, but it's actually a potential source of bugs. Such as:

- Giving users a direct handle to the index: depends on whether it's been allocated into simbody yet, which depends on `Component::extendRealizeTopology` being called. Potential footgun if downstream classes try to cache it too early. This can be fixed by providing callers with a wrapper (e.g. `CacheVariable`) that ensures the index is initialized before allowing a caller to do an index-related operation. This is effectively fixed by the range check that already happens in `CacheVariable`.

- Letting downstream classes store the handle: assuming they grab the handle at the correct time, or a correct wrapper is exposed (above), the next footgun is that the downstream class will want to store the resulting cached index locally. The risk here is that you can, logically speaking, end up with a situation where the downstream class owns a reference to a `CacheVariable` that has already been destructed by a call to `_namedCacheVariables.clear()` (as is done by `Component::reset`). That's basically a segfault/bizzare bug waiting to happen. 

  - One way to get around this is to just "make sure they don't use the wrappers after a call to `Component::reset()`, but that convention also depends on downstream implementers also knowing that the `std::map` is also cleared by `Component::clearConnections`, `Component::finalizeFromProperties`, and `Component::updateFromXMLNode`. 

  - Another way to get around this is to have the `std::map` store `shared_ptr<CacheVariable>`, so that the reference lifetime is extended beyond when the `CacheVariable`s are cleared from the underlying `std::map`. This would remove the possibility of a segfault, but introduce the possibility of `CacheVariable`s outliving their intended lifespan. To get around that, the `reset()` method should *also* reset all `CacheVariable`'s `index` arguments to `SimTK::NaN`s or whatever, so that using the `CacheVariable` beyond its intended lifetime causes a runtime exception that prompts downstream implementors to fix their lifetime semantics

- Letting users copy `Model` implementations: If a user copies/`.clone()`s a `Model` implementation then all of `Model`'s member properties are also going to be copied - including the `CacheVariable`s. Conventionally, the copied `Model` will call `reset()` or something to ensure that the copy performs a fresh initialization for `SimTK` etc. However, again, this is merely a convention. What we definitely don't want to happen is that the copy ends up with the ability to mutate the source's `CacheVariable`s, because the copied instance blindly copied references to the source's `CacheVariable`s. This is something that will currently be caught by the `std::map` lookup failing, but that failure will not happen if the lookup is skipped. Copying a `Model` implementation should, by default, reset all `CacheVariable` members to a senteniel (e.g. a `NaN`) so that using the copied `CacheVariable` without initializing it in the copy causes an appropriate explosion.