@adamkewley 2020-06-17
So the latest commit tries to introduce `CacheVariable<T>`, which is a type-safe wrapper around the underlying cache variable that will throw if it hasn't been allocated into SimTK yet. Users can request the `CacheVariable<T>` and then perform mutations, etc. on it. I've added some usage points into the codebase, and it seems like things are marginally faster (because the lookups have been reduced slightly, but not entirely eliminated):

repeats = 64
|                  Test Name | master [secs] | σ [secs] | branch [secs] | σ [secs] | Speedup |
| -------------------------- | ------------- | -------- | ------------- | -------- | ------- |
|                   Gait2354 |          0.27 |     0.01 |          0.27 |     0.00 |    1.02 |
|             ToyDropLanding |         15.74 |     0.26 |         15.13 |     0.16 |    1.04 |
|   ToyDropLanding_nomuscles |          0.47 |     0.00 |          0.46 |     0.00 |    1.00 |
|            passive_dynamic |          5.35 |     0.02 |          5.29 |     0.03 |    1.01 |
| passive_dynamic_noanalysis |          3.35 |     0.01 |          3.33 |     0.02 |    1.01 |
|                      Arm26 |          0.32 |     0.00 |          0.31 |     0.00 |    1.01 |
|             RajagopalModel |         17.59 |     0.09 |         17.30 |     0.07 |    1.02 |

This is a kind of middle-ground to where we want things to be. In this approach, users can request the `CacheVariable<T>` locally and then use it. The "ideal" approach would be for the downstream classes to hold onto the `CacheVariable<T>` as class members, so that they can skip lookups altogether. However, this is difficult because:

- I definitely don't want to support default initialization for `CacheVariable<T>` yet (it would need some `if` checks on the public API to ensure that it throws appropriately)
- I definitely don't want to support copy initialization for `CacheVariable<T>` yet (it could lead to aliasing of the underlying variable)
- I don't want to support half-initialized `CacheVariable<T>`s, but the creation of the cache variable (via `addCacheVariable`) has no access to the necessary `SimTK::State`.

Those two things could be dealt with by having a copy constructor that invalidates the copy (e.g. by nulling pointers then throwing whenever it's used) and by enforcing that `CacheVariable<T>`s are initialized using (e.g.) a constructor initializer. 

Overall, this isn't an automatically easy thing to fix because the creation point for cache variables (`addCacheVariable`) is separate from the actual creation point (topology realization) and the destruction point is either whenever the underlying storage is `clear`ed or the component destructs. Solutions suggesting just storing the `SimTK::` indexes could work, but are also reliant on knowing the lifetime semantics (the existing index getter methods can return stale indices).