@adamkewley 2020-06-18
Perf measurements show mostly a little noise, but some minor speedups in the `ToyDropLanding` (which is the model that instigated this investigation, and might use some of the usage points I've optimized):

```c++
repeats = 64
|                  Test Name | master [secs] | σ [secs] | branch [secs] | σ [secs] | Speedup |
| -------------------------- | ------------- | -------- | ------------- | -------- | ------- |
|                   Gait2354 |          0.26 |     0.01 |          0.26 |     0.00 |    1.01 |
|             ToyDropLanding |         14.78 |     0.06 |         14.01 |     0.06 |    1.05 |
|   ToyDropLanding_nomuscles |          0.44 |     0.00 |          0.43 |     0.00 |    1.01 |
|            passive_dynamic |          4.87 |     0.01 |          4.90 |     0.01 |    0.99 |
| passive_dynamic_noanalysis |          3.09 |     0.01 |          3.11 |     0.01 |    0.99 |
|                      Arm26 |          0.30 |     0.00 |          0.29 |     0.00 |    1.02 |
|             RajagopalModel |         15.93 |     0.05 |         15.19 |     0.11 |    1.05 |
```