@adamkewley 2020-06-19
So this latest implementation seems to work and speeds things up:

repeats = 64
|                  Test Name | master [secs] | σ [secs] | branch [secs] | σ [secs] | Speedup |
| -------------------------- | ------------- | -------- | ------------- | -------- | ------- |
|                   Gait2354 |          0.26 |     0.00 |          0.26 |     0.00 |    1.00 |
|             ToyDropLanding |         14.71 |     0.01 |         12.86 |     0.01 |    1.14 |
|   ToyDropLanding_nomuscles |          0.43 |     0.00 |          0.43 |     0.00 |    1.00 |
|            passive_dynamic |          4.87 |     0.00 |          4.86 |     0.00 |    1.00 |
| passive_dynamic_noanalysis |          3.07 |     0.00 |          3.06 |     0.00 |    1.00 |
|                      Arm26 |          0.29 |     0.00 |          0.28 |     0.00 |    1.06 |
|             RajagopalModel |         14.91 |     0.01 |         14.02 |     0.01 |    1.06 |

Missing full rollout (there's still some parts of the codebase using the "old" method) and testing, but it looks like a fairly solid speedup on my desktop, at least.